@@ -31,21 +31,36 @@
     "@electron-forge/plugin-fuses": "^7.8.1",
     "@electron-forge/plugin-webpack": "^7.8.1",
     "@electron/fuses": "^1.8.0",
-    "@eslint/js": "^9.30.1",
+    "@eslint/js": "^9.32.0",
+    "@stylistic/eslint-plugin": "^5.2.2",
+    "@types/electron-squirrel-startup": "^1.0.2",
+    "@types/lodash": "^4.17.20",
+    "@types/react": "^19.1.8",
+    "@types/react-dom": "^19.1.6",
+    "@typescript-eslint/parser": "^8.38.0",
     "@vercel/webpack-asset-relocator-loader": "^1.7.3",
     "css-loader": "^6.11.0",
     "electron": "37.2.1",
-    "eslint": "^9.30.1",
+    "eslint": "^9.32.0",
+    "eslint-plugin-import": "^2.32.0",
     "eslint-plugin-react": "^7.37.5",
+    "eslint-plugin-react-hooks": "^5.2.0",
     "globals": "^16.3.0",
+    "jiti": "^2.5.1",
     "node-loader": "^2.1.0",
+    "postcss": "^8.5.6",
+    "postcss-loader": "^8.1.1",
     "style-loader": "^3.3.4",
-    "typescript-eslint": "^8.36.0"
+    "typescript": "^5.8.3",
+    "typescript-eslint": "^8.38.0"
   },
   "dependencies": {
     "@tailwindcss/postcss": "^4.1.11",
     "electron-squirrel-startup": "^1.0.1",
-    "postcss": "^8.5.6",
-    "tailwindcss": "^4.1.11"
+    "lodash": "^4.17.21",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
+    "tailwindcss": "^4.1.11",
+    "ts-loader": "^9.5.2"
   }
 }
