@@ -1,4 +1,6 @@
-module.exports = [
+import type { ModuleOptions } from 'webpack';
+
+export const rules: Required<ModuleOptions>['rules'] = [
   // Add support for native node modules
   {
     // We're specifying native_modules in the test because the asset relocator loader generates a
@@ -13,6 +15,17 @@ module.exports = [
       loader: '@vercel/webpack-asset-relocator-loader',
       options: {
         outputAssetBase: 'native_modules',
+      },
+    },
+  },
+  {
+    test: /\.tsx?$/,
+    exclude: /node_modules/,
+    use: {
+      loader: 'ts-loader',
+      options: {
+        transpileOnly: true,
+        configFile: 'tsconfig.json',
       },
     },
   },
